chore: move types to pass e2e test that dont like import from scene
