+++ src/services/filters.ts
@@ -2,6 +2,11 @@ import { SelectableValue } from '@grafana/data';
 import { DetectedLabel } from './fields';
 import { ALL_VARIABLE_VALUE, LEVEL_VARIABLE_VALUE } from './variables';
 
+export enum FilterOp {
+  Equal = '=',
+  NotEqual = '!=',
+}
+
 // We want to show labels with cardinality 1 at the end of the list because they are less useful
 // And then we want to sort by cardinality - from lowest to highest
 export function sortLabelsByCardinality(a: DetectedLabel, b: DetectedLabel) {
